setting is now a separate module
